+++ Kernel/Output/HTML/Layout.pm
@@ -147,6 +147,15 @@ sub new {
                     last LANGUAGE;
                 }
             }
+            if (!$Self->{UserLanguage}) {
+                for my $Language ( reverse sort keys %Data ) {
+                    # If Browser requests 'vi', also offer 'vi_VI' even though we don't have 'vi'
+                    if ( $Language =~ m/^$BrowserLang/smxi) {
+                        $Self->{UserLanguage} = $Language;
+                        last LANGUAGE;
+                    }
+                }
+            }
         }
         $Self->{UserLanguage} ||= $Self->{ConfigObject}->Get('DefaultLanguage') || 'en';
     }